Timeskip simulation (#33)
Upgrades the simulation to time-skip when there are idle periods in the
system.
Results in ~30x speedup over previous simulation implementation.

Uses tokio
[pause](https://docs.rs/tokio/latest/tokio/time/fn.pause.html) and
[advance](https://docs.rs/tokio/latest/tokio/time/fn.advance.html).
These are

All uses of non-Tokio Instant type for time comparison now result in
compile errors. And all queries to system time and clock must go through
the type abstractions. These abstractions are located in
`iluvatar_library/src/clock.rs`
call `now()` to get an tokio `Instant` representing the current time,
and `get_global_clock()` to get the clock use by the system, which is
set dynamically if running under a simulation.

This ensures all parts of the system see the same time, and see time
elapse together and correctly as tokio moves the clock ahead underheath

---------

Co-authored-by: Alex Fuerst <alfuerst@iu.edu>